+++ src/Farmer/Arm/Insights.fs
@@ -3,7 +3,20 @@ module Farmer.Arm.Insights
 
 open Farmer
 
-let components = ResourceType("Microsoft.Insights/components", "2014-04-01")
+let private createComponents version = ResourceType("Microsoft.Insights/components", version)
+/// Classic AI instance
+let components = createComponents "2014-04-01"
+/// Workspace-enabled AI instance
+let componentsWorkspace = createComponents "2020-02-02"
+
+/// The type of AI instance to create.
+type InstanceKind =
+    | Classic
+    | Workspace of ResourceId
+    member this.ResourceType =
+        match this with
+        | Classic -> components
+        | Workspace _ -> componentsWorkspace
 
 type Components =
     { Name : ResourceName
@@ -11,7 +24,9 @@ type Components =
       LinkedWebsite : ResourceName option
       DisableIpMasking : bool
       SamplingPercentage : int
-      Tags: Map<string,string> }
+      InstanceKind : InstanceKind
+      Tags: Map<string,string>
+      Dependencies : ResourceId Set }
     interface IArmResource with
         member this.ResourceId = components.resourceId this.Name
         member this.JsonModel =
@@ -19,16 +34,24 @@ type Components =
                 match this.LinkedWebsite with
                 | Some linkedWebsite -> this.Tags.Add($"[concat('hidden-link:', resourceGroup().id, '/providers/Microsoft.Web/sites/', '{linkedWebsite.Value}')]", "Resource")
                 | None -> this.Tags
-
-            {| components.Create(this.Name, this.Location, tags = tags) with
-                 kind = "web"
-                 properties =
-                  {| name = this.Name.Value
-                     Application_Type = "web"
-                     ApplicationId =
-                       match this.LinkedWebsite with
-                       | Some linkedWebsite -> linkedWebsite.Value
-                       | None -> null
-                     DisableIpMasking = this.DisableIpMasking
-                     SamplingPercentage = this.SamplingPercentage |}
+            {| this.InstanceKind.ResourceType.Create(this.Name, this.Location, this.Dependencies, tags) with
+                kind = "web"
+                properties =
+                    {|
+                        name = this.Name.Value
+                        Application_Type = "web"
+                        ApplicationId =
+                            match this.LinkedWebsite with
+                            | Some linkedWebsite -> linkedWebsite.Value
+                            | None -> null
+                        DisableIpMasking = this.DisableIpMasking
+                        SamplingPercentage = this.SamplingPercentage
+                        IngestionMode =
+                            match this.InstanceKind with
+                            | Workspace _ -> "LogAnalytics"
+                            | Classic -> null
+                        WorkspaceResourceId =
+                            match this.InstanceKind with
+                            | Workspace resourceId -> resourceId.Eval()
+                            | Classic -> null |}
             |}