+++ src/Farmer/Builders/Builders.WebApp.fs
@@ -161,27 +161,27 @@ type SlotBuilder() =
     member this.AddConnectionStrings(state, connectionStrings:string list) :SlotConfig =
         connectionStrings
         |> List.fold (fun state key -> this.AddConnectionString(state, key)) state
-        
+
     /// Add Allowed ip for ip security restrictions
-    [<CustomOperation "add_allowed_ip_restriction">] 
-    member _.AllowIp(state, name, cidr:IPAddressCidr) : SlotConfig = 
+    [<CustomOperation "add_allowed_ip_restriction">]
+    member _.AllowIp(state, name, cidr:IPAddressCidr) : SlotConfig =
         { state with IpSecurityRestrictions = state.IpSecurityRestrictions @ [IpSecurityRestriction.Create name cidr Allow] }
-    member this.AllowIp(state, name, ip:Net.IPAddress) : SlotConfig = 
+    member this.AllowIp(state, name, ip:Net.IPAddress) : SlotConfig =
         let cidr = { Address = ip; Prefix = 32 }
         this.AllowIp(state, name, cidr)
-    member this.AllowIp(state, name, ip:string) : SlotConfig = 
-        let cidr = { Address = Net.IPAddress.Parse ip; Prefix = 32 }
+    member this.AllowIp(state, name, ip:string) : SlotConfig =
+        let cidr = IPAddressCidr.parse ip
         this.AllowIp(state, name, cidr)
     /// Add Denied ip for ip security restrictions
-    [<CustomOperation "add_denied_ip_restriction">] 
-    member _.DenyIp(state, name, cidr:IPAddressCidr) : SlotConfig = 
+    [<CustomOperation "add_denied_ip_restriction">]
+    member _.DenyIp(state, name, cidr:IPAddressCidr) : SlotConfig =
         { state with IpSecurityRestrictions = state.IpSecurityRestrictions @ [IpSecurityRestriction.Create name cidr Deny] }
-    member this.DenyIp(state, name, ip:Net.IPAddress) : SlotConfig = 
+    member this.DenyIp(state, name, ip:Net.IPAddress) : SlotConfig =
         let cidr = { Address = ip; Prefix = 32 }
         this.DenyIp(state, name, cidr)
-    member this.DenyIp(state, name, ip:string) : SlotConfig = 
-        let cidr = { Address = Net.IPAddress.Parse ip; Prefix = 32 }
-        this.DenyIp(state, name, cidr) 
+    member this.DenyIp(state, name, ip:string) : SlotConfig =
+        let cidr = IPAddressCidr.parse ip
+        this.DenyIp(state, name, cidr)
     interface ITaggable<SlotConfig> with member _.Add state tags = { state with Tags = state.Tags |> Map.merge tags }
     interface IDependable<SlotConfig> with member _.Add state newDeps = { state with Dependencies = state.Dependencies + newDeps }
 
@@ -232,7 +232,7 @@ type WebAppConfig =
       AutomaticLoggingExtension : bool
       SiteExtensions : ExtensionName Set
       PrivateEndpoints: (LinkedResource * string option) Set
-      CustomDomains : Map<string,DomainConfig> 
+      CustomDomains : Map<string,DomainConfig>
       DockerPort: int option
       ZoneRedundant : FeatureFlag option }
     member this.Name = this.CommonWebConfig.Name
@@ -315,9 +315,9 @@ type WebAppConfig =
                     | Linux, Some _
                     | _ , None ->
                         ()
-                        
+
                     yield! this.DockerPort |> Option.mapList AppSettings.WebsitesPort
-                    
+
                     if this.DockerCi then "DOCKER_ENABLE_CI", "true"
                 ]
 
@@ -484,6 +484,8 @@ type WebAppConfig =
                   Location = location
                   DisableIpMasking = false
                   SamplingPercentage = 100
+                  InstanceKind = Classic
+                  Dependencies = Set.empty
                   LinkedWebsite =
                     match this.CommonWebConfig.OperatingSystem with
                     | Windows -> Some this.Name.ResourceName
@@ -503,7 +505,7 @@ type WebAppConfig =
                   MaximumElasticWorkerCount = this.MaximumElasticWorkerCount
                   OperatingSystem = this.CommonWebConfig.OperatingSystem
                   ZoneRedundant = this.ZoneRedundant
-                  Tags = this.Tags}
+                  Tags = this.Tags }
             | _ ->
                 ()
 
@@ -556,7 +558,7 @@ type WebAppConfig =
                           DomainName = customDomain }
 
                     // Get the resource group which contains the app service plan
-                    let aspRgName = 
+                    let aspRgName =
                       match this.CommonWebConfig.ServicePlan with
                       | LinkedResource linked -> linked.ResourceId.ResourceGroup
                       | _ -> None
@@ -566,7 +568,7 @@ type WebAppConfig =
                     // To keep the code simple, I opted to always nest the certificate deployment. - TheRSP 2021-12-14
                     let certificateDeployment = resourceGroup { 
                         name (aspRgName |> Option.defaultValue "[resourceGroup().name]")
-                        add_resource 
+                        add_resource
                           { cert with
                               SiteId = Unmanaged cert.SiteId.ResourceId
                               ServicePlanId = Unmanaged cert.ServicePlanId.ResourceId }
@@ -861,16 +863,19 @@ module Extensions =
             { current with ConnectionStrings = current.ConnectionStrings.Add(key, (ParameterSetting (SecureParameter key), Custom)) }
             |> this.Wrap state
         member this.AddConnectionString(state:'T, (key, value:ArmExpression)) =
+            this.AddConnectionString(state, (key, value, Custom))
+        member this.AddConnectionString(state:'T, (key, value:ArmExpression, kind)) =
             let current = this.Get state
-            { current with ConnectionStrings = current.ConnectionStrings.Add(key, (ExpressionSetting value, Custom)) }
+            { current with ConnectionStrings = current.ConnectionStrings.Add(key, (ExpressionSetting value, kind)) }
             |> this.Wrap state
+            
         /// Creates a set of connection strings of the web app whose values will be supplied as secret parameters.
         [<CustomOperation "connection_strings">]
         member this.AddConnectionStrings(state:'T, connectionStrings) =
             let current = this.Get state
             connectionStrings
             |> List.fold (fun (state:CommonWebConfig) (key, value:ArmExpression) ->
-               { state with ConnectionStrings = state.ConnectionStrings.Add(key, (ExpressionSetting value, Custom)) }) current 
+               { state with ConnectionStrings = state.ConnectionStrings.Add(key, (ExpressionSetting value, Custom)) }) current
             |> this.Wrap state
         /// Sets an app setting of the web app in the form "key" "value".
         [<CustomOperation "add_identity">]
@@ -1001,12 +1006,15 @@ module Extensions =
             |> List.fold (fun (state:CommonWebConfig) (key, value: string) -> { state with Settings = state.Settings.Add(key, LiteralSetting value); SlotSettingNames = state.SlotSettingNames.Add(key) }) current
             |> this.Wrap state
         /// Add Allowed ip for ip security restrictions
-        [<CustomOperation "add_allowed_ip_restriction">] 
-        member this.AllowIp(state:'T, name, ip) = 
+        [<CustomOperation "add_allowed_ip_restriction">]
+        member this.AllowIp(state:'T, name, ip:IPAddressCidr) =
+            this.Map state (fun x -> { x with IpSecurityRestrictions = IpSecurityRestriction.Create name ip Allow :: x.IpSecurityRestrictions })
+        member this.AllowIp(state:'T, name, ip:string) =
+            let ip = IPAddressCidr.parse ip
             this.Map state (fun x -> { x with IpSecurityRestrictions = IpSecurityRestriction.Create name ip Allow :: x.IpSecurityRestrictions })
         /// Add Denied ip for ip security restrictions
-        [<CustomOperation "add_denied_ip_restriction">] 
-        member this.DenyIp(state:'T, name, ip) = 
+        [<CustomOperation "add_denied_ip_restriction">]
+        member this.DenyIp(state:'T, name, ip) =
             this.Map state (fun x -> { x with IpSecurityRestrictions = IpSecurityRestriction.Create name ip Deny :: x.IpSecurityRestrictions })
         /// Integrate this app with a virtual network subnet
         [<CustomOperation "route_via_vnet">]